Config translator and explainer model (#39)
Enable `shell` and `cli` to config model that is used for translation
and explanation.
Use command `@config model set <kind> <model>` to set the model for kind
= translation or explanation
Use command `@config model show [<kind>]` to show the current set model
(and the URL for the model)

Also add command to enable/disable an agent (i.e. both translation and
action for an agent) using `@config agent <agentname>`