Introduce transient app agents (#81)
- Sub agent like paliodb, crosswords, etc. only is active if the web
page is opened. Before, it just toggles the agent directory, overwriting
the user setting. Introduce a runtime state for transient agent to set
whether it is active or not, separate from whether the user has
enable/disable them.
- remove the need to expose "currentTranslatorName" to the agent's
session context. (One less "NYI" to run separate process)
- Also add a lock to make sure we don't change the state while we are
servicing a request int he toggle.
- Fix the `currentTranslatorName` getting overwritten on every request.
The state is meant to record what is the last translator use so we will
try to resolve the next request first.
- Improve the "source" agent for status messages, it shouldn't get bound
to the current translator's name (or dispatcher). Now it defaults to
dispatcher, but get passed in during translation using the translator
that is currently trying.
- Fix the bug where the metric disappear on dynamic displays.