Reducing HTML size further (#82)
- Added logic to remove HTML nodes that are not visible (checking
element.visible property). Also removed nodes such as "cookieJar" that
appear often on ecommerce sites but do not affect layout/visuals.
- Enable websocket clients to specify an ID as a query parameter when
connecting. Updated the service to use this parameter to dedupe
connections when clients occasionally end up with multiple connections
to the service.
- Brought back use of embeddings for browser tab title matching. This
time the model code is in the agent rather than the browser extension.
- Update messageType names for browser and desktop actions