Remove impression interpreter (#114)
Remove ImpressionInterpreter because it doesn't work across process
boundary when agent is running out of proc.
Since it has minimal use right now, just change it to precompute the
value that is needed for the dispatcher (instead of delay calling to
interpret). We can design something different if there are additional
need.