Shell: switch processShellRequest from invoke to just message response/reply (#115)
In the render process the order of message processing send from the main
process and invoke completion can be out of order.
That make some of the display message for action comes after the
`processShellRequest` is finished.
In the tts case, that means the content is not spoken.

Change from using `invoke` to just message, so that (in theory and
manual testing) that the completion of `processShellRequest` will be in
the same order as other message.

Also
- fix the case where the browser voices get loaded after the setting
view is construction.
- chat view get the tts when the settings are loaded.