Fix streaming (#117)
PR #112 introduce a regression for streaming. That is because the
"content" field of the first chunk is "undefined", and so the code end
up returning `<multimodel content>`.

The non-streaming `complete` function has a laxer requirement for the
response, which skips any text before the first "{" and after the last
"}" thus avoiding the error (probably a hold over when LLM JSON mode
doesn't exist). The incremental JSON is more strict and bailed on trying
parse, and broke streaming.

Looking at both OpenAI and Azure API reference, unlike the input, the
response messages are not a complete `PromptSection`. The content for
the response can only be string currently.

The change introduce more accurate typing for the response, and revert
the logic change from PR#112