Fix for 'Type quit OR exit to exit' which omits the @ prefix (#260)
At the start of a session, there's a message `Type quit OR exit to
exit`.
However, if you type `quit`, it's an error.
The reason is that this is a command prefixed with `@` (though handled
specially in the command parser).

My fix changes this to `Type @quit OR @exit to exit`,
taking the configured command prefix and list of stopcommands into
account.

While I was here, I also fixed a missing quote in the final welcome
message (`'hello`).