Plaintext to commands in code chat (#262)
This is my first magnum opus. :-)

When you type plaintext in the code chat example, it now attempts to
transform it to an existing command.
It does this by constructing a TypeChat schema (once) for the relevant
`NamedArgs` from the handlers' metadata.
When the plaintext transformation succeeds to produce valid JSON
(guaranteed by TypeChat),
the corresponding command is invoked. Hope for the best. We're not using
history here.

I should mention that this required some changes to `interactiveApp.ts`.
I think those are backwards compatible, at least with this version I
didn't have to fix any other examples or packages.
(It is a little bit finicky; I had to turn `CommandHandler` into a
union.