Explainer: reduce entity reference confusion error rate to 0 on test data (#311)
Explainer: 
- Explainer should use the request instead of the conversation history
as the source of the property value if they are present in both. This
change reduces confusion:
- Add the user request and translate at the top of the property
explanation so that LLM know what we are talking about specifically
before the schema.
-  Detect if an entity property value and the substring from the request
is the same, indicating that it should be a property coming from the
request instead of from history. Ask the LLM a correction the
explanation.
- Adjust comment to make it clearer that we prefer if the substring is
from the request.
- Additionall fix couple of places where string comparison should be
case insensitive. GPT-4o model has a slightly higher likelihood to
generate substrings/sub-phrase with different cases

Infra:
- Add `--entities` flag to `cli data regenerate` to simulate history
context with entities that are also in the request by synthesizing the
history with property value.
- Explainer will not capture the reference error with the explanation as
a non-correctable correction, so that it can be captured in the test
data generation