First release of Spelunker (#313)
Right now this mostly exists for debugging the loading and querying of
Python code.

The src/main.ts program takes 0 or more .py files on the command line
and imports them all into the database (/data/spelunker). The Python
code is parsed and split into "chunks" (which may be nested) by a Python
script src/chunker.py. (Parsing Python is easy using the `ast` stdlib
module; the cost we pay is that we have an silent external dependency on
`python3`.)

The Python script writes a JSON list of chunks to stdout which is read
by src/pythonChunker.ts and reconstituted into a list of `Chunk` objects
in the TS program. (The Python and TypeScript versions of this class and
the `Blob` class must be kept in sync manually.)

The src/pythonImporter.ts file then takes this list of chunks and writes
them to an `ObjectFolder` (one chunk per folder item), as well as
storing an embedding for it using a `SemanticCodeIndex`. Since I had
lots of issues here, I verbosely log the comments created by the
embedding code interspersed with the source code lines. (This currently
shows the documenter is way to eager to create a comment for each line;
I'd rather have a higher-level summary for each chunk. Working on that
next.)

Once all files have been processed, `main()` then sits in a loop asking
the user for a line of text representing a query and looking the nearest
N neighbors and shows those hits. (N is currently 5; ideally we should
be able reduce that to 1 or 2.) The UI is currently crap.

---------

Co-authored-by: Guido van Rossum <gvanrossum@microsoft.com>