New approach to Spelunker embeddings (#321)
We create a new code reviewer that produces a single summary paragraph
per chunk. We save these summaries in a separate object folder. This is
faster and produces more useful embeddings.

Also two speedups:
- Chunkify all files together in a single chunker.py invocation
- Parallelize processing of all chunks for a single file

Finally, some UI improvements: the summaries are not printed during
import, but they are printed together on retrieval. This makes importing
much quieter now that we've debugged it more.

This required a small interface change to codeIndex.ts: instead of
taking a CodeReviewer, createSemanticCodeIndex now takes a narrower
interface that only has a document() method.