Adding setWallpaper action to the Desktop Agent.
It has a hacky shortcut where it assumes a generated.png image in the user_files folder, instead of figuring out the filePath value from the data in the context.

The image agent's image generation code was also updated to create a generated.png file every time it generates a new image, along with some other plumbing code to funnel the storage location where needed. All that is also part of the shortcut and shouldn't be needed.