Fix translator override (#333)
The current method of extending the TypeChat translator to include
history context for request prompt creation, and streaming callback on
each translation invocation, makes the translator unable to handle
multiple call at the same time. Now that we use the agent translators to
detect context usage for the explainer with will cause overlapping calls
and might cause the override to go across the call and cause random
errors.

Change the method of extending by:
- For agent translation, expand request prompt with chat history upfront
using a wrapped `TypeAgentTranslator` before passing it to the TypeChat
translator.
- For the streaming, it needs to pass the incremental parser to the
model, and the only way to do so thru TypeChat is hiding it in the
prompt premble.

Additional changes:
- Remove `ChatMessage` types for image sections, where TypeChat
`PromptSection` already have those now.
- Trace the translation result (and renamed the prompt trace name to be
consistent `typeagent:translation:<typename>:prompt` and
`typeagent:translation:<typename>:result`