Add NPR Dataset to Python Directory (#335)
Code and materials for the generation of a dataset encompassing NPR's
All Things Considered Podcast. Also contains code for working with a
test episode of the Behind the Tech podcast. Additional files allow for
testing the dataset using RAG and comparing against Structured RAG
outputs.

The NPR dataset encompasses all podcasts in the All Things Considered
Archive after 2011. This includes:

- 713 Episodes
- 11125 Sections
- 223726 Turns

Further details about the dataset are included in the `README.md` file. 

The `generate_chunks.py` file contains code for generating chunks from
this dataset mechanically using dialogue turns as chunks. This directory
is self contained and does not modify files anywhere else in the repo.