Fixing image url handling code in the setWallpaper so it works with urls that have query parameters (#351)
Fixing image url handling code in the setWallpaper so it works with urls
that have query parameters. This change also created unique names for
the downloaded file, otherwise the code later in this method can pick up
a different file while resolving relative paths (e.g. it was picking up
a generated_00.png file from another session, as generated image urls
follow that name convention).

There is also an unrelated change to tokenCounter to log token
increments in addition to the totals. This makes it easier to track the
token impact of individual calls.