[spelunker] Add nearestNeighborsPairs to TextTable and createTextIndex (#380)
This extends the TextTable interface and its two known implementations
to support nearestNeighborsPairs, which returns scored pairs of (entry
text, array of source ids) for the nearest matches.

I needed this for detailed logging what was going in. (I found that
"ast" is somehow a universal attractor in embedding space!)

Also add some clarifying (for me!) comments to createTextIndex (in
knowledgeIndex.js) and added some extra (redundant) types that made the
meaning of various variables clearer to me (I was mightily confused by
the two uses of strings, for entry values and for internal Ids created
by the semantic index's put() method).

And some cleanup in spelunker (notably removing a special case to avoid
double-scoring exact matches that was based on a misunderstanding) and
of course use the new nearestNeighborsPairs for better logging of the
index matching mechanism.

I refrained from more serious refactoring.