Generated action schema (#420)
Use our schema parser and parse the composed schema into memory. Use
that to generate the schema sent to the LLM (for now in the exact form)
and use the in-memory representation for validation (avoid using
typescript compiler for it)

Generation is on by default, and can be turned off using `@config
translation schema generation off`

- Add action schema parsing of in-memory buffer
- Add `ActionSchemaFile` to keep track of action schemas within a single
source file.