Switch inline switcher and multiple action schema to use dynamic action schema creation API (#424)
Inline switcher and multiple action schema to use dynamic action schema
creation API, whether we use the in-memory validator or type script
validate.

Also, leveraging the above implemented a function to generate composite
schema that is sent to the LLM (that includes the injected, inline
switcher and multiple action) and wired that to `agent-cli schema
--generated`. It is not used in the actual action translation yet, and
will be in a follow up PR.

Additional change:
- Add capability to have type reference that is not resolved (for the
multiple action case), with the assumption that the definition will be
somewhere.
- During generation, it will error out by default if there are
unresolved type reference, but will ignore if { strict: false} is passed
in to the generation function.
- Add template parameter for SchemaType types so that it can narrow the
type of fields, etc. Use to specify the shape of an `ActionSchema`, and
has the compiler/intelligence know of the fields (`actionName` and
`parameters`)
- Starting to transition the `translator` nomenclature to refer to a set
of action schema for an app agent, and switch to `action` or
`actionSchema`
- Change the natural order of schema generation to be "depth" first so
that type references are close to each other.