Improve console output for CLI (#448)
- Reduce concept and removed requestIO and using clientIO and the
command context directly.
- Forcing adjustment of the cursor on every input key to the cursor
problem when certain agent's emoji is displayed in the console prompt,
and node's r`eadline` didn't account for fullwidth of certain emoji
character.
- externalize tracking of history so that `askYesNo` input not be part
of the "history" by not sharing `readline`
  - Reduce lifetime of `readline` to just one question
instance.
- Allow Ctrl-C to work to kill the process during action by set `stdin`
to raw mode and resume the stream after `readline.question` so that we
can drain and continue to listen to input.
- Refactor and move console related code to `helpers/console` in for the
dispatcher.

Also:
- Allow boolean value "true" and "false" to be specified as value for
boolean flag (previously, having the flag means true, so no way to
specify false to turn the flag off)