Switch to turn off initial schema pick using embedding. (#490)
- `@config translation switch embedding [on|off]` to toggle the initial
schema pick using embedding
- `@config translation switch fix <agentName>` command to turn off
switching and set the schema to use.
- Fix the problem of missing injected schema (including `unknown`) when
inline switcher is off,
- Split the clarify schema into 3 to better drive change of thought for
each case for the clarification question. This also help make sure it
doesn't overly generalized and generate clarification when it should
just go to `unknown` action.
- Add additional message to the unknown action to indicate if the likely
action for the request (using embedding) is not active, or if the
switching of action is completely off.
- Improve fetch error by pulling out the inner exception's message (e.g.
Connection time out)