API Storage Provider Support (#539)
Abstracted the Azure Blob storage code in `api` into
`TypeAgentStorageProvider` to support using s3 or any other storage
provider for session storage as well.

Note: I do not have the ability to test the Azure Storage Provider,
would be prudent to have someone with azure credentials do a test run
with blob storage to ensure it is working as expected.

---------

Co-authored-by: Steve Lucco <steveluc@users.noreply.github.com>