[Spelunker] rewrite as an agent (#547)
This uses a different algorithm for selecting context -- instead of
pre-computing a set of indexes (which takes minutes) it now uses an
approach similar to John Lam's "ask" example (not open source). That is,
we partition the full list of chunks into up to `MAX_CONCURRENCY`
sub-lists, and pass these concurrently to a fast, cheap model (GPT 4.0
mini) which is prompted to select the chunks most relevant to the user
query, with a "relevance" score between 0 and 1. We then take the 30
highest-scoring chunks overall, and send those with the user query to a
smarter model (GPT 4.0.2).

It's still Python only. I use the same chunker.py script to chunkify
Python files.

This doesn't work as well as the original Spelunker example, but it's
useful to compare the two algorithms, and I expect to evolve this into
something better (possibly the same algorithm as before, possibly
something hybrid, possibly something new).

Some specific differences:
- No indexing required, which means no startup cost
- Chunks are not summarized to English text
- No conversation history (I need to figure out how to get this)
- No error handling or exponential backoff

How to use:
- Start shell or cli interactive
- Type `Set spelunker focus to <folder-name>` (also try `get spelunker
focus`)
- Type `@config request spelunker`
- Type away, asking questions about `<folder-name>`
- To leave exclusive mode, use `@config request dispatcher`
- You can also use `@spelunker <question>` without `@config request`

In my experience, with a code base of ~300-400 chunks, the context
selection takes between 9-15 seconds, and the smarter model takes
between 4-6 seconds. That's much slower than the original Spelunker
(whose first phase only takes 1-2 seconds), but it's still usable.

Anecdotally I've seen more elaborate answers given by the agent -- not
sure if the answer was actually *better*.