Improve embedding heuristics and pared downed prompt context (#556)
Fix the scenario: 
- `add the ingredients for chess pie to my grocery list`
- `don't need the sugar`  


Couple of issues:
1. When picking the initial schema group to use, both request's
embedding matches the calendar actions because the calendar schema
description with the embedding is included in the embedding, and the
action itself doesn't have description. That diluted the focus and let
it match much different range of requests.
2. Even if the initial pick is wrong, the actual translation prompt ends
up choosing the "removeEvent" action, instead of switching to calendar.
The issue here is that when we pare down the schema to only the top 5
likely actions, it loses the context that these are calendar actions.
  
The changes are:
1. Remove descriptions from the action embedding. (Fixes 1 and 2).
2. Add comments to the calendar actions. (Fixes 1 and 2)
3. Regenerate the type alias for the schema group for a union of the
pared down actions and add the description of the schema group along
with it to give it context (Fixes #2).

With these fixes, the initial pick correctly choose the `list` agent.
Also even when `calendar` was picked (without fix 1 and 2), it will be
able switch back to the `list` agent instead of get stuck with calendar
agent.