[Spelunker] Try to work with entities (#594)
Includes an update to the chunker to return the function/class/module
name as a new field (codeName).

The entity work is incomplete, I produce entities, and they end up in
the session history, but I am having a hard time having them passed back
into spelunker. At least they can be used by the translator sometimes.

Example of what works:
- "spelunker describe tracing in more detail": produces a long bullet
list, and some entities
- "spelunker describe the first two bullets in more detail" : makes two
separate calls to Spelunker:
- "Describe the initialization and loading in the ConfigurationReader
class in more detail."
- "Describe the flattening and merging in the ConfigurationReader class
in more detail."
 
It is passing stuff to each of the last two spelunker calls, but it uses
the filename/lineno instead of the unique chunk ID that I want.

I will leave finishing this for another day, but I want this set of
changes merged since it does work to some extent.