Fix race condition when running preload script for inline browser (#619)
Our preload script for the inline browser is built as an ESM module.
There is a known issue when loading ESM preload scripts:
https://github.com/electron/electron/issues/40777. This causes the
preload script to sometimes run after the hosted page has been rendered.
When this happens, the hosted page cannot access the APIs exposed by the
preload script, leading to unexpected behavior. The fix is to update the
build to generate this preload script as a commonJS module.