Pending action support in multiple actions (#621)
If an action requires the content of the result from previous action,
multiple action will generate a pending request entry.
The pending request entry is converted into a "PendingRequestAction",
and during execute, the request will be translated again with the
current history context.

- A lot of the diff is from breaking apart the file
`requestCommandHandler.ts`
- `executeActions` is not a queue to allow pending request action to be
expanded to more actions.
- streaming now allow different `actionIndex` as the streaming target so
that translation for pending request can be streamed too, but if the
translation is not a multiple action.
- Add toggles for result id and pending for multiple actions.