[Spelunker] Fix mini model endpoint (#650)
The endpoint name for the mini model was misspelled, so we quietly got
the default model (gpt-4-o-2).

After fixing it, I noticed that the summarizing and selecting phases
(which use the mini model) became significantly *slower*. (There's a lot
of variability in the measurements but the difference was pretty
consistent).

So now the fix is to set the mini model endpoint name to `undefined`,
which means it's the same old default model, but there's no misleading
misspelled endpoint name, and a comment explaining the reason.