Better error handling, retry logic, logging, refactoring
- Refactor getAllSourceFiles to return a list of (file, mtime, size) records.
- Use lstat instead of stat, and skip all symlinks.
- Refactor loadDatabase accordingly.

Focusing the spelunker on all of TypeAgent/ts (about 90,000 LoC),
we now see fewer errors, and explicit logging of the errors that do occur.
It takes about 80 seconds to summarize that code, and another 80
to select relevant chunks.

Alas, we have to stop using the summaries,
because they make the prompt too large.
(TODO: Don't compute summaries if there are too many chunks.)