Entity with uniqueId roundtrip (#666)
- If the agent returns entities with uniqueId, and the same "name" is
used in the following action, we will put the entity in the EntityMap.
Only entity with uniqueId from the same agent will be populated.
- For the result id reference, change the entity map to include only
result that are from the same agent and referenced in the action.
- Change how we determine uniqueness of entities based on name and type
only. Attach a `v<num>` suffix to entities that have same name and type
to distinguish each other. Therefore, entity with the same name and type
but different uniqueId will have different suffix.
- LLM may change lower/upper case or add/remove diacritical. When lookup
the action result with the entity, use a normalized form that remove
those difference for look up.
- Future investigate of whether we need to do more to associate the
entities with the original conversation so that the LLM will know about
the context and make better choices
- Player agent now populate the entity with id and make use of them when
returned.