[Spelunker] Adjust breadcrumbs using summaries (#693)
The intention is to make breadcrumbs embedded in a class, module or
outer function more informative even when the chunk which it represents
is not included in the context.

This applies to chunk selection and oracle requests; not to summarize
requests (or we'd have a chicken/egg problem).

For example, a breadcrumb of the form
```ts
    function beginRefresh ...
```
is replaced with
```ts
     // Prevents multiple calls to refresh while one is pending.
    function beginRefresh(): Promise<Result<string>> ...
```
Note how the indentation is preserved.

Also made many logging tweaks and one small schema tweak.