Add function calling support (#701)
- Add function calling support (in addition to structure output). They
are very similar except for a little bit different between the schemas,
the REST call param and the output location (note that streaming is not
implemented yet)
- Fix JSON schema with $ref subschema, where description is not
supported (that means that we will miss comments for fields with a type
reference).
- When Json schema (either structured output or function calling) is
enabled, don't inject the typescript schema from the main prompt and
also don't validate the output (assuming the service have validated them
already since we have strict mode on).
- Also with `cli test translate` allow number of repeat to change if we
process all the data from the `-i` input if specified.
- Fix the incorrect check if history is enabled/disabled or not.