[Spelunker] Tweak eval tooling (#743)
Log successful spelunker requests in `~/.spelunker.log`. (A place where
it isn't easily accidentally erased.)
Add a simple tool to read back `~/.spelunker.log`. (For manual
inspection, for now.)
Add a simple script to dump the salient parts of an eval database to a
text file `eval.sql`. (Tool to read it back is TBD.)
Rename IdType to ChunkId in `chunker.py`.
Preserve `Questions` and `Scores` tables when redoing the eval setup.
Lint.