Combine all schema when all agent switching methods are off (#769)
- If embedding, inline and search switching are off, we will try to
combine all the active schema into one single prompt.
- For now, we will error out if there are type name or action name that
conflicts between the schemas.
- Require the translator keep track of the `actionName` -> `schemaName`
mapping. The steps to translate the request will keep track of the
translator the action is from so that it can resolve the schema name
after finalizing. (This will allow us to do mix and match later).

Additionally:
- Refactor TypeAgentTranslator construction to be more flexible by
passing in the inject/switch action config instead of the state to
compute it internally.
- Skip the translation optimization of using embedding to choose a
subset of action to prompt first when we retry the same schema during
the translation in the case where LLM want to look at other actions from
the same schema, removing the need to do an explicit pass without
optimization on multiple actions.
- Fix the @ command error output to use the default color instead of
black which could lead to "black on black".