browser: fix inconsistent automation errors (#799)
- the extension relied on `chrome.windows.getCurrent()` to get the
current window in view. However, when the devTools view is undocked, it
is treated as the active window. This leads to errors where the service
worker tries to send events to the devtools view instead of the content
window.
- There was an error where some code paths tried to access
chrome.storage apis in the content script, causing the script to crash.