Use lookupAndAnswer to self-clarify unresolve references. (#852)
Initial implementation of using `lookupAndAnswer` to
clarifyUnresolvedReference`

- When there is an unresolve references `clarifyUnresolvedReference`,
try to use `lookupAndAnswer` to find the answer to the reference and
re-translate. This includes references to previous conversation or
internet facts.
- Add capability to an actions' result to return a set of actions to be
queue up to executed, with the policy that actions needs to be in the
same app agent (except for dispatcher). This allows any actions to chain
actions together. Used by `clarifyUnresolvedReference` to return the
actions with the information resolved using `lookupAndAnswer`.
- Move `lookupAndAnswer` to the dispatcher. (Future plan is to have the
information source extensible by other agents, and the implementation
can move out of the dispatcher when it provides an extension point)

Tested with `cli test translate` with no change in stability. Added e2e
test for conversation lookup, but internet lookup is not tested yet
(will be in future PR)