Serialization and writing to file works!
In the end, the various serialization() methods
return podcast_embeddings as numpy arrays,
either a single 1536-long vector, for a single
embedding, or a 2D array with each row being a
1536-float32-long vector for multiple embeddings.

Only the ultimate serialization data
(ConversationFileData.binaryData) is a bytearray.
We could even optimize that by writing the binary
data to a file as we go.

Next up: deserialization and reading from files.