[python/knowpro] Implement serialization! (#880)
Deserialization is next.

I *think* the schema is the same as for the TypeScript version, but I'd
love to see that validated.

I went back and forth a lot over how to represent embeddings in the JSON
(before they are extracted into the binary data file). In the end I kept
pointing to the original vectorbase contents (i.e., an
NDArray[np.float32] or None), accumulating them into a bytearray at the
very end (in def write_conversation_data_to_file()).

There are still some helper types in serialization.py that I'm not sure
I want to keep, but for now I went with the flow.

The lack of `?.` in Python (PEP 505) is really making the code ugly.