Example MCP integration. (#894)
Added MCP file system agent (`mcpfilesystem`) as an example of auto MCP
integration to TypeAgent.
- The agent `mcpfilesystem` is not enabled by default. Use `@config
agent mcpfilesystem` to enable
- On first run, enabling schema and action will show an error because
the allowed directories need to be provided. Use the command
`@mcpfilesystem server <directory>...` to configure. A restart is needed
afterwards.

Details:
- Implement Tools JSON schema converter to our parsed action schema
format, which can generate TS or back to JSON schema for prompts.
- Implemented MCP agent provider to automatically wrap a stdio MCP
server. (can be extend to other transport in the future.

Other changes: 
- Allow schema parse error to be non-fatal at start up (but disable the
schema)
- Shell will catch and display dispatcher initialization error and exit
(instead of silently hang).
- Fix Shell's demo command `@shell run`. Need to send the
send-input-text to the chatView instead of mainWindow.
- Clean up unused tracking of injected action names.