Fix test instability (#1006)
Get occasional failure in `sqliteCommon.ts` createDatabase function in
test complaining unable to create database because the path doesn't
exist. That's because in `ensureDir`, it calls the promise version of
`mkdir` without await. So it doesn't guarantee that the directory is
actually created before it returns. Switch to the `sync` version