Add script to drive the Azure service deployment process (#1172)
Require Azure CLI, that is already logged in with `az login` and default
subscription set.

In the `tools/scripts` directory:
- Create: run `node azureDeploy.mjs create` to deploy the azure service.
- Delete: run `node azureDeploy.mjs delete` to completely delete the
deployed azure service.
- Purge: run `node azureDeploy.mjs purge` to purge deleted services (if
`--purge false` is specified in delete`)

After deploy, it will automatically runs `getKeys` to populate `.env`
file.

Change the getKeys to also save the keyvault name so that it will use
that as the default if not specified.