Track activity context in chat history and enable replay with activity information. (#1234)
Save activity context in the chat history and set it during replay via
importing.

- Hide the "entries" and provide API on `ChatHistory` for operations.  
- - Add import/export of chat history to use the "input" format, change
`@history save` and `@history insert`.
- Refactor the code that updates the activities context from action
result and reuse that when importing chat history.
- - Store the schema name as the source of the assistant entries in the
chat history so that we can use it to fill out the schema name of the
end action during import.
- Add `any` support for schema creator and validator so that we can
validate the "any" fields in the activity context when reading from
exported chat history from file.