[python/knowpro] Queries including a time range work! (#1242)
There was a one-line bugfix in searchlang.py (compile_when() wasn't
returning what it had computed).

The rest was implementing missing things in query.py and search.py.
(There's still more missing, but not exercised much.)

Oh, and one tweak to interfaces.py: Fix the repr() of DateRange to
enclose the arguments in string quotes, so pretty_print() can handle it.

---------

Co-authored-by: Guido van Rossum <gvanrossum@microsoft.com>