[python/knowpro] Finish MatchMessagesBooleanExpr subclasses; add automatic evaluation script (#1250)
`test/cmpsearch.py` now doesn't just show you the queries to run; it
runs the queries and compares the actual answers to the expected ones
using a dot product between embeddings.

Run `python test/cmpsearch.py` to try it. This assumes a checkout of the
(internal) AISystems-Archive repo exists as a sibling of the TypeAgent
repo.

Various small tweaks elsewhere.