Fix `tsc` incremental build in browser and markdown agent (#1261)
- Remove the unnecessary `tsBuildInfoFile` location. Currently
`fluid-build` doesn't support specifying custom `tsBuildInfoFile`.
location. Fluid repro [PR
pending](https://github.com/microsoft/FluidFramework/pull/24911)
- Also, this fix a problem where two `tsc` command output to the same
`tsBuildInfo` location.
- Remove duplicate `src/view/shared` files in browser agents
`tsconfig.json` in `src/view/server`, resulting mismatch of `tsc`
thinking those files as input, but not included in the `tsBuildInfo`.
`fluid-build` detected that and trigger the build, as those could be
"new" files for the build that needs to be process.
- Markdown agent's `src/view/site` 
- must be `composite` for `incremental` to be enabled in `tsc` (which
`fuild-build` relies on).
     - Fix typing errro after `composite` is enabled.
- Also the `vite` steps afterward for that directory overwrite the
result, and thus the `tsBuildInfo` is deleted, causing `fluid-build` to
always rebuild. Change it to `noEmit` and change the output directory
different then the `vite` step.